Fix typo and move helper to utils.go
