+++ neo4j-cli/aura/internal/subcommands/dataapi/graphql/corspolicy/allowedorigin/allowed_origin.go
@@ -1,14 +1,7 @@
 package allowedorigin
 
 import (
-	"encoding/json"
-	"fmt"
-	"net/http"
-	"os"
-
 	"github.com/neo4j/cli/common/clicfg"
-	"github.com/neo4j/cli/common/clierr"
-	"github.com/neo4j/cli/neo4j-cli/aura/internal/api"
 	"github.com/spf13/cobra"
 )
 
@@ -22,41 +15,4 @@ func NewCmd(cfg *clicfg.Config) *cobra.Command {
 	cmd.AddCommand(NewRemoveCmd(cfg))
 
 	return cmd
-}
-
-type DetailedBody struct {
-	Data DataBody `json:"data"`
-}
-
-type DataBody struct {
-	Security SecurityBody `json:"security"`
-}
-
-type SecurityBody struct {
-	CorsPolicy CorsPolicyBody `json:"cors_policy"`
-}
-
-type CorsPolicyBody struct {
-	AllowedOrigins []string `json:"allowed_origins"`
-}
-
-func getGetExistingOrigins(cfg *clicfg.Config, dataApiId, instanceId string) ([]string, error) {
-	getPath := fmt.Sprintf("/instances/%s/data-apis/graphql/%s", instanceId, dataApiId)
-	getResBody, statusCode, err := api.MakeRequest(cfg, getPath, &api.RequestConfig{
-		Method: http.MethodGet,
-	})
-	if err != nil {
-		return nil, err
-	}
-	if statusCode != http.StatusOK {
-		panic(clierr.NewFatalError("unexpected status code %d running CLI with args %s, please report an issue in https://github.com/neo4j/cli", statusCode, os.Args[1:]))
-	}
-
-	var parsedGetResBody DetailedBody
-	err = json.Unmarshal(getResBody, &parsedGetResBody)
-	if err != nil {
-		panic(err)
-	}
-
-	return parsedGetResBody.Data.Security.CorsPolicy.AllowedOrigins, nil
 }
